Added error handling using an error interceptor, material dialog, and protecting user from accessing APIs that would have been caught in the backend, just for a user-friendly approach.

+++ backend/routes/posts.js
@@ -42,8 +42,10 @@ router.post('',
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: host + '/images/' + req.file.filename
+        imagePath: host + '/images/' + req.file.filename,
+        creator: req.userData.userId
     });
+
     post.save(post).then(createdPost => {
         res.status(201).json({
             message: 'Post added sucessfully',
@@ -52,6 +54,11 @@ router.post('',
                 id: createdPost._id,
             }
         })
+    })
+    .catch(error => {
+        res.status(500).json({
+            message: "Creating a post failed!"
+        })
     });
 });
 
@@ -59,22 +66,38 @@ router.put('/:id',
     checkAuth,
     multer({storage: storage}).single('image'),
     (req, res, next) => {
-    let imagePath = req.body.imagePath;
-    if (req.file) {
-        const host = req.protocol + '://' + req.get('host');
-        imagePath = host + '/images/' + req.file.filename;
+        let imagePath = req.body.imagePath;
+        if (req.file) {
+            const host = req.protocol + '://' + req.get('host');
+            imagePath = host + '/images/' + req.file.filename;
+        }
+        const post = new Post({
+            _id: req.body.id,
+            title: req.body.title,
+            content: req.body.content,
+            imagePath: imagePath,
+            creator: req.userData.userId
+        });
+
+        // Post.updateOne will only update a post that has the correct "_id" and "creator"
+        // So cool - we let mongoose protect changes to the post for only the creator of the post.
+        Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post)
+            .then((result) => {
+                // Might or might not have found a valid post
+                if (result.nModified > 0) {
+                    res.status(200).json({message: "Post update sucessful!"});
+                } else {
+                    res.status(401).json({message: "Not authorized to update post!"});
+                }
+            })
+            .catch(error => {
+                // This is a technical error, such as a database connetion error
+                res.status(500).toJSON({
+                    message: "Couldn't update post!"
+                })
+            })
     }
-    const post = new Post({
-        _id: req.body.id,
-        title: req.body.title,
-        content: req.body.content,
-        imagePath: imagePath
-    });
-    Post.updateOne({_id: req.params.id}, post)
-        .then((result) => {
-            res.status(200).json({message: "Update sucessful"});
-        })
-});
+);
 
 router.get('', (req, res, next) => {
     const pageSize = +req.query.pageSize;
@@ -99,26 +122,49 @@ router.get('', (req, res, next) => {
                 maxPosts: count
             });
         })
+        .catch(error => {
+            res.status(500).json({
+                message: "Fetching posts failed!"
+            })
+        })
 });
 
 router.get('/:id', (req, res, next) => {
     Post.findById(req.params.id)
         .then(post => {
+            // A post was either found or not found.
             if (post) {
                 res.status(200).json(post);
             } else {
                 res.status(404).json({message: 'Post not found!'});
             }
         })
+        .catch(error => {
+            // This error is a technical error.
+            res.status(500).json({
+                message: "Fetching posts failed!"
+            })
+        })
 });
 
 router.delete('/:id',
     checkAuth,
     (req, res, next) => {
-    Post.deleteOne({_id: req.params.id})
-        .then((result) => {
-            res.status(200).json({message: `Post ${req.params.id} deleted!`});
-        })
-});
+        Post.deleteOne({_id: req.params.id, creator: req.userData.userId})
+            .then((result) => {
+                if (result.n > 0) {
+                    res.status(200).json({message: `Post ${req.params.id} deleted!`});
+                } else {
+                    res.status(401).json({message: "Not authorized to delete post"});
+                }
+            })
+            .catch(error => {
+                // This error is a technical error.
+                res.status(500).json({
+                    message: "Deleting post failed!"
+                })
+            })
+    }
+);
 
 module.exports = router;
